requirements.txt
PyExecJS
beautifulsoup4
requests